+++ packages/roosterjs-content-model/lib/formatHandlers/TableCellFormatHandler.ts
@@ -1,23 +1,11 @@
 import { backgroundColorFormatHandler } from './common/backgroundColorFormatHandler';
 import { borderFormatHandler } from './common/borderFormatHandler';
-import { createBooleanDefinition, createObjectDefinition } from 'roosterjs-editor-dom';
-import { createMetadataFormatHandler } from './common/metadataFormatHandler';
+import { ContentModelTableCellFormat } from '../publicTypes/format/ContentModelTableCellFormat';
 import { FormatHandler } from './FormatHandler';
 import { sizeFormatHandler } from './common/sizeFormatHandler';
+import { tableCellMetadataFormatHandler } from './table/tableCellMetadataFormatHandler';
 import { textAlignFormatHandler } from './common/textAlignFormatHandler';
 import { verticalAlignFormatHandler } from './common/verticalAlignFormatHandler';
-import {
-    ContentModelTableCellFormat,
-    TableCellFormat,
-} from '../publicTypes/format/ContentModelTableCellFormat';
-
-const TableCellFormatDefinition = createObjectDefinition<Required<TableCellFormat>>(
-    {
-        bgColorOverride: createBooleanDefinition(true /** isOptional */),
-    },
-    false /* isOptional */,
-    true /** allowNull */
-);
 
 /**
  * @internal
@@ -28,5 +16,5 @@ export const TableCellFormatHandlers: FormatHandler<ContentModelTableCellFormat>
     backgroundColorFormatHandler,
     textAlignFormatHandler,
     verticalAlignFormatHandler,
-    createMetadataFormatHandler<TableCellFormat>(TableCellFormatDefinition),
+    tableCellMetadataFormatHandler,
 ];
