+++ packages/roosterjs-content-model/lib/formatHandlers/utils/createMetadataFormatHandler.ts
